Moved StacKArray to a proper folder
